+++ wallet/src/account/types/address_serde.rs
@@ -6,12 +6,12 @@ use serde::{de::Visitor, Deserializer, Serializer};
 use crate::account::types::address::AddressWrapper;
 
 /// custom AddressWrapperserialization to use the bech32 representation
-pub fn serialize<S: Serializer>(address: &AddressWrapper, s: S) -> std::result::Result<S::Ok, S::Error> {
+pub(crate) fn serialize<S: Serializer>(address: &AddressWrapper, s: S) -> std::result::Result<S::Ok, S::Error> {
     s.serialize_str(&address.to_bech32())
 }
 
 /// custom AddressWrapper derialization to use the bech32 representation
-pub fn deserialize<'de, D>(deserializer: D) -> Result<AddressWrapper, D::Error>
+pub(crate) fn deserialize<'de, D>(deserializer: D) -> Result<AddressWrapper, D::Error>
 where
     D: Deserializer<'de>,
 {